refactor: extract parseAsanaUrl into src/lib/asana-url
Move the Asana URL parser out of the WebFetch hook into src/lib/asana-url.ts so
the CLI can reuse it. The hook now imports parseAsanaUrl; behavior is unchanged.
Parser tests move to test/lib/asana-url.test.ts; the hook test keeps only the
hook-specific buildCliGuidance/decideForToolCall cases.
